pyproject.toml
[project]
name = "ai-examples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.10.0",
    "litellm>=1.75.5.post1",
]
